Bug Id: 389\n The default value for carrier detect wait is now 0 (turned off on the editor panels)

+++ yukon-client/common/src/com/cannontech/database/db/port/PortSettings.java
@@ -10,7 +10,7 @@ public class PortSettings extends DBPersistent
 {
 	private Integer portID = null;
 	private Integer baudRate = new Integer(1200);
-	private Integer cdWait = new Integer(1000);
+	private Integer cdWait = new Integer(0);
 	private String  lineSettings = "8N1";
 
 /**
